Move test data to its own directory
